pyproject.toml
[tool.poetry]
name = "oeleo"
version = "0.1.0"
description = ""
authors = ["jepegit <jepe@ife.no>"]

[tool.poetry.dependencies]
python = "^3.10"
peewee = "^3.15.0"
python-dotenv = "^0.20.0"
rich = "^12.4.4"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
